requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarl==1.9.2